+++ lib/database/database.dart
@@ -1,11 +1,14 @@
 /**
  * @file database.dart
- * @brief Floor 数据库配置
+ * @brief 应用数据库定义
  *
- * 配置所有实体及 DAO 接口。
+ * 定义应用程序数据库结构，包含各模块所需的表和DAO。
  */
 
 import 'dart:async';
+import 'package:floor/floor.dart';
+import 'package:sqflite/sqflite.dart' as sqflite;
+
 import '../models/event_item.dart';
 import '../models/customer_item.dart';
 import '../models/expense_item.dart';
@@ -15,12 +18,12 @@ import '../dao/customer_dao.dart';
 import '../dao/expense_dao.dart';
 import '../dao/maintenance_dao.dart';
 
-part 'database.g.dart';
+part 'database.g.dart'; // 这行很重要，不要遗漏
 
 @Database(version: 1, entities: [EventItem, CustomerItem, ExpenseItem, MaintenanceItem])
 abstract class AppDatabase extends FloorDatabase {
   EventDao get eventDao;
-  CustomerDao get customerDao;
+  //CustomerDao get customerDao;
   ExpenseDao get expenseDao;
   MaintenanceDao get maintenanceDao;
 }
