+++ lib/models/event_item.dart
@@ -1,21 +1,21 @@
 /**
  * @file event_item.dart
- * @brief 事件规划模块实体类
+ * @brief 事件数据模型
  *
- * 定义事件记录的各个字段。
+ * 定义事件的数据结构
  */
 
-import 'package:flutter/material.dart';
 import 'package:floor/floor.dart';
+
 @entity
 class EventItem {
   @PrimaryKey(autoGenerate: true)
   final int? id;
-  final String name;
+
+  final String title;
   final String date;
-  final String time;
   final String location;
   final String description;
 
-  EventItem(this.id, this.name, this.date, this.time, this.location, this.description);
+  EventItem(this.id, this.title, this.date, this.location, this.description);
 }
