+++ lib/models/maintenance_item.dart
@@ -1,8 +1,8 @@
 /**
  * @file maintenance_item.dart
- * @brief 车辆维护模块实体类
+ * @brief 车辆维护数据模型
  *
- * 定义车辆维护记录的各个字段。
+ * 定义车辆维护记录的数据结构
  */
 
 import 'package:floor/floor.dart';
@@ -11,13 +11,12 @@ import 'package:floor/floor.dart';
 class MaintenanceItem {
   @PrimaryKey(autoGenerate: true)
   final int? id;
-  final String vehicleName;
-  final String vehicleType;
-  final String serviceType;
-  final String serviceDate;
-  final String mileage;
-  final String cost;
 
-  MaintenanceItem(this.id, this.vehicleName, this.vehicleType, this.serviceType,
-      this.serviceDate, this.mileage, this.cost);
+  final String vehicleName;
+  final String maintenanceType;
+  final String date;
+  final String cost;
+  final String notes;
+
+  MaintenanceItem(this.id, this.vehicleName, this.maintenanceType, this.date, this.cost, this.notes);
 }
